TinyMCE: Class Selection Options is now Element Options
This allows you to configure more than just classes to be added and
removed as options in the TinyMCE Options dropdown. Now you can
configure any HTML attribute, which is both powerful and dangerous.

Also moved TinyMCE Cloud API Key to an environment variable and removed
spellchecker as a default option (it's not free).