Update objects after editing their fields
After updating a field, on blur replace the object with updated HTML
that was generated as a result of saving the object. Only do this when
another field in that object is not in focus, to prevent editors being
rebuilt while a user is trying to use them.

Updating the HTML in this way allows backend changes to reflect in the
frontend without a page refresh. This could be template logic based on
values, or shortcode processing, or anything else that might change
based on content.