Better object re-rendering and in-DOM replacement
Now rather than rendering the object on its own, re-render the whole
page and pluck the object from the response. This means it will render
with full page context which some templates require.

To support this, a UID strategy has been employed to ensure the frontend
can pluck the correct object from the response. This also allows it to
much more efficiently search the page for other versions of the same
object.